requirements.txt
pygame
mysql-connector-python

